Use floats instead of doubles

+++ ConsoleGame/Game.cpp
@@ -72,15 +72,15 @@ Direction Game::GetPlayerDirection() const
 
 bool Game::IsPlayerMoving() const
 {
-   return _player->GetVelocityX() != 0. || _player->GetVelocityY() != 0.;
+   return _player->GetVelocityX() != 0 || _player->GetVelocityY() != 0;
 }
 
-double Game::GetArenaWidth() const
+float Game::GetArenaWidth() const
 {
    return _gameDefs->ArenaDefs->Width;
 }
 
-double Game::GetArenaHeight() const
+float Game::GetArenaHeight() const
 {
    return _gameDefs->ArenaDefs->Height;
 }
@@ -103,27 +103,27 @@ void Game::MovePlayer()
 {
    _arenaPlayerPositionX += ( _player->GetVelocityX() / _gameDefs->FramesPerSecond );
 
-   if ( _arenaPlayerPositionX < 0. )
+   if ( _arenaPlayerPositionX < 0 )
    {
-      _arenaPlayerPositionX = 0.;
+      _arenaPlayerPositionX = 0;
       _player->StopX();
    }
    else if ( _arenaPlayerPositionX >= _gameDefs->ArenaDefs->Width )
    {
-      _arenaPlayerPositionX = _gameDefs->ArenaDefs->Width - 1.;
+      _arenaPlayerPositionX = _gameDefs->ArenaDefs->Width - 1;
       _player->StopX();
    }
 
    _arenaPlayerPositionY += ( _player->GetVelocityY() / _gameDefs->FramesPerSecond );
 
-   if ( _arenaPlayerPositionY < 0. )
+   if ( _arenaPlayerPositionY < 0 )
    {
-      _arenaPlayerPositionY = 0.;
+      _arenaPlayerPositionY = 0;
       _player->StopY();
    }
    else if ( _arenaPlayerPositionY >= _gameDefs->ArenaDefs->Height )
    {
-      _arenaPlayerPositionY = _gameDefs->ArenaDefs->Height - 1.;
+      _arenaPlayerPositionY = _gameDefs->ArenaDefs->Height - 1;
       _player->StopY();
    }
 }
